Move PM specific support to their own PM prefixed components
