+++ components/src/PMPagination/index.js
@@ -1,0 +1,1 @@
+export { PMPagination } from "./PMPagination";
