+++ components/src/PMTable/index.js
@@ -1,0 +1,1 @@
+export { PMTable } from "./PMTable";
